services/rag_server/pyproject.toml
[project]
name = "rag-server"
version = "0.1.0"
description = "RAG server for document processing and querying"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.3",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.2",
    "python-multipart>=0.0.20",
    "llama-index-core>=0.14.4",
    "llama-index-readers-docling>=0.4.1",
    "llama-index-node-parser-docling>=0.1.0",
    "llama-index-embeddings-ollama>=0.8.3",
    "llama-index-llms-ollama>=0.7.4",
    "llama-index-llms-openai>=0.3.0",
    "llama-index-llms-anthropic>=0.6.0",
    "llama-index-llms-google-genai>=0.4.0",
    "llama-index-llms-deepseek>=0.2.2",
    "llama-index-vector-stores-postgres>=0.3.4",
    "llama-index-postprocessor-sbert-rerank>=0.4.0",
    "docling>=2.53.0",
    "sse-starlette>=3.0.2",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.52.2",
    # PostgreSQL stack
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "pgvector>=0.3.6",
    "alembic>=1.14.0",
    "tembo-pgmq-python>=1.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "fpdf2>=2.8.0",
]
bench = [
    "datasets>=3.2.0",
    "streamlit>=1.41.0",
    "pandas>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "eval: RAG evaluation tests (require --run-eval flag and ANTHROPIC_API_KEY)",
    "integration: Integration tests requiring docker services (postgres, ollama)",
    "slow: Tests taking > 30s",
]
addopts = "--strict-markers"
